Fix clippy's hard error and the remaining lint warnings
`cargo clippy` had been failing outright, so it wasn't usable as a gate:
RAMTAS's `'scan: loop` never actually loops — the inner loop either breaks
out of it or spins forever — which clippy rejects as an error rather than a
warning. A labelled *block* says the same thing truthfully.

The rest are small: a redundant `u16` cast, a one-arm `match` that reads
better as an `if`, and an `#[allow]` on `Op::len` with a note that an opcode
length is not a collection length, so there is deliberately no `is_empty`.

`cargo clippy --all-targets` is now clean and worth keeping that way.

+++ crates/mos6502/src/cpu.rs
@@ -378,7 +378,7 @@ impl Cpu {
             "RTS" => {
                 let lo = self.pull(bus) as u16;
                 let hi = self.pull(bus) as u16;
-                self.pc = ((lo | (hi << 8)).wrapping_add(1)) as u16;
+                self.pc = (lo | (hi << 8)).wrapping_add(1);
                 6
             }
             "RTI" => {
